% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCol.R
\name{rep.col}
\alias{rep.col}
\title{Two functions to duplicate columns}
\usage{
\method{rep}{col}(x, n)
}
\arguments{
\item{x}{column to be repeated}

\item{n}{number of times to be repeated}
}
\value{
matrix with x repeated n times
}
\description{
These functions allow easy replication of rows and columns
}
\examples{
na
rep.col()
}
\seealso{
na
}

