% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repRow.R
\name{rep.row}
\alias{rep.row}
\title{Two functions to duplicate rows and columns}
\usage{
\method{rep}{row}(x, n)
}
\arguments{
\item{x}{row to be repeated}

\item{n}{number of times to be repeated}
}
\value{
matrix with x repeated n times
}
\description{
These functions allow easy replication of rows and columns
}
\examples{
na
rep.row()
}
\seealso{
na
}

