% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repCol.R
\name{repCol}
\alias{repCol}
\title{Two functions to duplicate columns}
\usage{
repCol(x, n)
}
\arguments{
\item{x}{column to be repeated}

\item{n}{number of times to be repeated}
}
\value{
matrix with x repeated n times
}
\description{
These functions allow easy replication of rows and columns
}
\examples{
repCol(x,n)
}
\seealso{
https://www.r-bloggers.com/a-quick-way-to-do-row-repeat-and-col-repeat-rep-row-rep-col/ and is written by TszKin Julian
}

