% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repColrepRow.R
\name{repColrepRow}
\alias{repColrepRow}
\alias{repRow}
\alias{repCol}
\title{Two functions to duplicate rows and columns}
\usage{
repRow(x, n)

repCol(x, n)
}
\arguments{
\item{x}{row/column to be repeated}

\item{n}{number of times to be repeated}
}
\value{
matrix with x repeated n times
}
\description{
These functions allow easy replication of rows and columns
}
\examples{
\dontrun{repRow(x,n)}
}
\seealso{
https://www.r-bloggers.com/a-quick-way-to-do-row-repeat-and-col-repeat-rep-row-rep-col/ and is written by TszKin Julian
}

